Add SearchBar to result page
